Improve permissions window + color & title support for action button
I modified the font weight & colors of permission item so the titles can stand out more.

I also added bolder titles for the applications. While this is not ideal right now because there's too much bold titles in the window, I believe it's essential for the app names to be in bold. It's more important than the categories while you browse the repo so it should stand out more. Maybe reduce the font size of category & remove uppercases, idk.

I also tried to make the install button more obvious by giving it a name and a new background color for adw-gtk3 themes. Everything was all the same color in the current version, so I believe guiding the user by adding more colors is good.

However, I didn't succeed to change font color, and on both themes, its color appears too similar to the background. Sorry if it's not good code, it's my first contribution and honestly I know almost anything of programming other than CSS, so I did my best to understand the code and the GTK wiki.